+++ frontend/src/components/SenatorPortrait/index.css
