+++ frontend/src/pages/GamePage/index.css
